Adiconando conexão com o banco e os métodos

+++ Crud_Login/services/IServiceUsuario.cs
@@ -9,10 +9,10 @@ namespace Crud_Login.services {
     internal interface IServiceUsuario {
 
         void salvarUsuario(Usuario usuario);
-        Usuario getUsuarioPorId(long id);
-        bool atualizarUsuario(long id, Usuario usuario);
-        bool existePorId(long id);
-        bool deletarUsuario(long id);
+        Usuario getUsuarioPorId(int id);
+        bool atualizarUsuario(int id, Usuario usuario);
+        bool existePorId(int id);
+        bool deletarUsuario(int id);
 
     }
 }
